extract skr resource sync
