+++ internal/manifest/skrresources/resource_converter.go
@@ -1,4 +1,4 @@
-package v2
+package skrresources
 
 import (
 	"errors"
