+++ internal/manifest/skrresources/ssa.go
@@ -1,4 +1,4 @@
-package v2
+package skrresources
 
 import (
 	"context"
